+++ tests/functional/test_wps_package.py
@@ -41,7 +41,6 @@ from weaver.formats import (
     IANA_NAMESPACE,
     get_cwl_file_format
 )
-
 from weaver.processes.constants import CWL_REQUIREMENT_APP_DOCKER, CWL_REQUIREMENT_INIT_WORKDIR
 from weaver.utils import get_any_value